Handling configurations as separated exports
