+++ packages/commons/src/config/httpServerConfig.ts
@@ -1,0 +1,18 @@
+import { z } from "zod";
+import { APIEndpoint } from "../model/apiEndpoint.js";
+
+export const HTTPServerConfig = z
+  .object({
+    HOST: APIEndpoint,
+    PORT: z.coerce.number().min(1001),
+  })
+  .transform((c) => ({
+    host: c.HOST,
+    port: c.PORT,
+  }));
+export type HTTPServerConfig = z.infer<typeof HTTPServerConfig>;
+
+export type CommonConfig = z.infer<typeof HTTPServerConfig>;
+
+export const httpServerConfig: () => HTTPServerConfig = () =>
+HTTPServerConfig.parse(process.env);
