+++ packages/commons/src/config/index.ts
@@ -1,3 +1,2 @@
-
-export * from "./awsConfig.js";
-export * from "./commonConfig.js";
+export { AWSConfig } from "./awsConfig.js";
+export { httpServerConfig, HTTPServerConfig } from "./httpServerConfig.js";
