+++ packages/commons/src/config/loggerConfig.ts
@@ -1,0 +1,14 @@
+import { z } from "zod";
+
+export const LoggerConfig = z
+  .object({
+    LOG_LEVEL: z.enum(["debug", "info", "warn", "error"]),
+  })
+  .transform((c) => ({
+    logLevel: c.LOG_LEVEL,
+  }));
+
+export type LoggerConfig = z.infer<typeof LoggerConfig>;
+
+export const loggerConfig: () => LoggerConfig = () =>
+  LoggerConfig.parse(process.env);
