+++ packages/commons/src/index.ts
@@ -3,3 +3,4 @@ export * from "./context/index.js";
 export * from "./logging/index.js";
 export * from "./model/apiEndpoint.js";
 export * from "./types/index.js";
+export * from "./config/index.js";
